返回用户id

+++ aries-base/src/main/java/com/cf/aries/base/dao/UserInfoMapper.java
@@ -17,7 +17,7 @@ public interface UserInfoMapper {
 
     int insert(UserInfo record);
 
-    int insertSelective(UserInfo record);
+    long insertSelective(UserInfo record);
 
     List<UserInfo> selectByExample(UserInfoExample example);
